[Feat] 파티 편성칸 UI 작업 및 캐릭터 프리팹 변경

+++ Assets/Workspace/HSD/Scripts/UI/CharacterList_UI.cs
@@ -38,7 +38,8 @@ public class CharacterList_UI : MonoBehaviour
 
     public void GradeSorting()
     {
-        System.Array.Sort(_synergyUnits, (a, b) => b.Status.Data.Grade.CompareTo(a.Status.Data.Grade));
+        System.Array.Sort(_synergyUnits, (a, b) =>
+            ((int)b.Status.Data.Grade).CompareTo((int)a.Status.Data.Grade));
 
         for (int i = 0; i < _synergyUnits.Length; i++)
         {
